requirements.txt
uvicorn
fastapi~=0.65.1
passlib~=1.7.4
bcrypt
psycopg2
sqlalchemy~=1.4.17
alembic~=1.6.5
PyJWT~=2.1.0
pydantic~=1.8.2